requirements.txt
azure-ai-inference
python-dotenv
azure-core
tiktoken
fastapi
uvicorn
pydantic
python-multipart